refactor: thinning
